pyproject.toml
[project]
name = "agentsafe"
version = "0.1.0"
description = "Lightweight AI agent safety wrapper with observability, powered by NVIDIA Nemotron"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "agentsafe contributors" }]
keywords = ["ai-safety", "agents", "nemotron", "guardrails", "observability", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
openai = ["openai>=1.0.0"]
ui = [
    "gradio>=6.9.0",
]
dashboard = ["textual>=0.70.0"]
all = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "openai>=1.0.0",
    "gradio>=6.9.0",
    "textual>=0.70.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.urls]
Homepage = "https://github.com/agentsafe/agentsafe"
Repository = "https://github.com/agentsafe/agentsafe"

[project.scripts]
agentsafe = "agentsafe.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentsafe"]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
